requirements.txt
aenum==3.1.16
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
algokit==2.10.2
algokit-utils==4.2.3
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.67.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
anywidget==0.9.21
argcomplete==3.6.3
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
babel==2.17.0
backports.tarfile==1.2.0
bases==0.3.0
beautifulsoup4==4.14.3
bitsandbytes==0.48.2
bleach==6.3.0
bqplot==0.12.45
branca==0.8.2
breizhcrops==0.0.4.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
cftime==1.6.5
charset-normalizer==3.4.4
click==8.1.8
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
copier==9.11.3
cryptography==44.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docstring-parser==0.14.1
dunamai==1.25.0
earthengine-api==1.7.3
eerepr==0.1.2
einops==0.8.1
exceptiongroup==1.3.1
executing==2.0.1
fastapi==0.135.3
fastjsonschema==2.21.2
filelock==3.19.1
fiona==1.10.1
folium==0.20.0
fonttools==4.61.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
funcy==2.0
future==1.0.0
geemap==0.36.6
geocoder==1.38.1
geojson==3.2.0
geopandas==1.1.1
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.2.1
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
ipyevents==2.0.4
ipyfilechooser==0.6.0
ipykernel==7.1.0
ipyleaflet==0.20.0
ipython==8.37.0
ipywidgets==8.1.8
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jedi==0.19.2
Jinja2==3.1.6
jinja2-ansible-filters==1.3.2
jiter==0.13.0
joblib==1.5.2
json5==0.12.1
jsonargparse==4.44.0
jsondiff==2.2.1
jsonnet==0.21.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-leaflet==0.20.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
keyring==25.6.0
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lark==1.3.1
lightly==1.5.22
lightly-utils==0.0.2
lightning==2.6.0
lightning-utilities==0.15.2
linkify-it-py==2.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mistune==3.1.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
mslex==1.3.0
multidict==6.7.0
multiformats==0.3.1
multiformats-config==0.3.1
narwhals==2.13.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.7.3
networkx==3.3
notebook==7.5.0
notebook_shim==0.2.4
numpy==2.2.6
omegaconf==2.3.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==1.0.4
pillow==12.0.0
pipx==1.8.0
platformdirs==4.5.0
plotly==6.5.0
pluggy==1.6.0
plumbum==1.10.0
prometheus_client==0.23.1
prompt_toolkit==3.0.51
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
psygnal==0.15.0
pure_eval==0.2.3
py-algorand-sdk==2.11.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclip==0.7.0
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
PyNaCl==1.6.2
pyogrio==0.12.1
pyparsing==3.2.5
pyperclip==1.11.0
pyproj==3.7.1
pyshp==3.0.3
pyteal==0.27.0
pytest==9.0.1
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytorch-lightning==2.6.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
questionary==1.10.0
rasterio==1.4.3
ratelim==0.1.6
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
scooby==0.11.0
seaborn==0.13.2
segmentation_models_pytorch==0.5.0
semantic-version==2.10.0
Send2Trash==1.8.3
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
terminado==0.18.1
textual==3.7.1
threadpoolctl==3.6.0
timm==1.0.22
tinycss2==1.4.0
tomli==2.3.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchgeo==0.6.2
torchmetrics==1.8.2
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.3
typer-slim==0.20.0
typeshed_client==2.8.2
typing-inspection==0.4.2
typing-validation==1.2.12
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
userpath==1.9.2
uvicorn==0.44.0
visdom==0.2.4
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
widgetsnbextension==4.0.15
xgboost==3.1.2
xyzservices==2025.11.0
yarl==1.22.0
zipp==3.23.0
